add serialization test and fix NumUniqueSeparators
