+++ nlp_primitives/num_unique_separators.py
@@ -3,24 +3,24 @@ from featuretools.primitives import TransformPrimitive
 from woodwork.column_schema import ColumnSchema
 from woodwork.logical_types import IntegerNullable, NaturalLanguage
 
-NATURAL_LANGUAGE_SEPARATORS = " .,!?;\n"
+NATURAL_LANGUAGE_SEPARATORS = [" ", ".", ",", "!", "?", ";", "\n"]
 
 
 class NumUniqueSeparators(TransformPrimitive):
     """Calculates the number of unique separators.
 
     Description:
-        Given a string and an iterable of separators, determine
+        Given a string and a list of separators, determine
         the number of unique separators in each string. If a string
         is null determined by pd.isnull return pd.NA.
 
     Args:
-        separators (str, optional): an iterable of characters to count.
-            " .,!?;\n" is used by default.
+        separators (list, optional): a list of separator characters to count.
+            `[`" ", ".", ",", "!", "?", ";", "\n"]` is used by default.
 
     Examples:
-        >>> x = ['First. Line.', 'This. is the second, line!', 'notinlist@#$%^%&']
-        >>> num_unique_separators = NumUniqueSeparators(".,!")
+        >>> x = ["First. Line.", "This. is the second, line!", "notinlist@#$%^%&"]
+        >>> num_unique_separators = NumUniqueSeparators([".", ",", "!"])
         >>> num_unique_separators(x).tolist()
         [1, 3, 0]
     """
@@ -31,13 +31,13 @@ class NumUniqueSeparators(TransformPrimitive):
 
     def __init__(self, separators=NATURAL_LANGUAGE_SEPARATORS):
         assert separators is not None, "separators needs to be defined"
-        self.separators = set(separators)
+        self.separators = separators
 
     def get_function(self):
         def count_unique_separator(s):
             if pd.isnull(s):
                 return pd.NA
-            return len(self.separators.intersection(set(s)))
+            return len(set(self.separators).intersection(set(s)))
 
         def get_separator_count(column):
             return column.apply(count_unique_separator)
