+++ nlp_primitives/tests/test_utils.py
@@ -8,7 +8,13 @@ except ImportError:
 
 import featuretools as ft
 import pytest
-from featuretools import dfs, list_primitives
+from featuretools import (
+    calculate_feature_matrix,
+    dfs,
+    list_primitives,
+    load_features,
+    save_features
+)
 from featuretools.tests.testing_utils import make_ecommerce_entityset
 
 ft.primitives._load_primitives()
@@ -49,6 +55,33 @@ class PrimitiveT:
         for name, parameter in init_params.parameters.items():
             if parameter.default is not parameter.empty:
                 assert hasattr(primitive_, name)
+
+    def test_serialize(self, es):
+        features = dfs(entityset=es,
+                       target_dataframe_name="log",
+                       trans_primitives=[self.primitive],
+                       max_features=-1,
+                       max_depth=3,
+                       features_only=True)
+
+        feat_to_serialize = None
+        for feature in features:
+            if feature.primitive.__class__ == self.primitive:
+                feat_to_serialize = feature
+                break
+            for base_feature in feature.get_dependencies(deep=True):
+                if base_feature.primitive.__class__ == self.primitive:
+                    feat_to_serialize = base_feature
+                    break
+        assert feat_to_serialize is not None
+
+        df1 = calculate_feature_matrix([feat_to_serialize], entityset=es)
+
+        new_feat = load_features(save_features([feat_to_serialize]))[0]
+
+        df2 = calculate_feature_matrix([new_feat], entityset=es)
+
+        assert df1.equals(df2)
 
 
 def find_applicable_primitives(primitive):
